Rename: CoreData Entity mapping method 파일 분리
