+++ AKCOO/Data/Service/CountryProfile+fromEntity.swift
@@ -1,21 +1,13 @@
 //
-//  CountryProfile+CoreData.swift
+//  CountryProfile+fromEntity.swift
 //  AKCOO
 //
-//  Created by Anjin on 11/22/24.
+//  Created by Anjin on 11/24/24.
 //
 
 import CoreData
 
 extension CountryProfile {
-  static func toEntity(context: NSManagedObjectContext, profile: CountryProfile) -> CountryProfileEntity {
-    let entity = CountryProfileEntity(context: context)
-    entity.name = profile.name
-    entity.currencyUnitTitle = profile.currency.unitTitle
-    entity.currencyExchangeRate = profile.currency.unit
-    return entity
-  }
-  
   static func fromEntity(entity: CountryProfileEntity) -> Result<CountryProfile?, Error> {
     guard
       let name = entity.name,
