+++ AKCOO/Data/Service/UserQuestion+fromEntity.swift
@@ -1,21 +1,13 @@
 //
-//  UserQuestion+CoreData.swift
+//  UserQuestion+fromEntity.swift
 //  AKCOO
 //
-//  Created by Anjin on 11/22/24.
+//  Created by Anjin on 11/24/24.
 //
 
 import CoreData
 
 extension UserQuestion {
-  static func toEntity(context: NSManagedObjectContext, question: UserQuestion) -> UserQuestionEntity {
-    let entity = UserQuestionEntity(context: context)
-    entity.country = question.country
-    entity.category = question.category
-    entity.amount = question.amount
-    return entity
-  }
-  
   static func fromEntity(entity: UserQuestionEntity) -> Result<UserQuestion?, Error> {
     guard
       let country = entity.country,
